speech-to-text/requirements.txt
livekit-agents>=0.12.15
livekit-plugins-aws>=0.1.0
livekit-plugins-azure>=0.5.4
livekit-plugins-google>=0.10.5
livekit-plugins-openai>=0.11.0
livekit-plugins-deepgram>=0.6.19
livekit-plugins-assemblyai>=0.2.2
livekit-plugins-fal>=0.2.3
livekit-plugins-silero>=0.7.4
livekit-plugins-clova @ git+https://github.com/livekit/agents#egg=livekit-plugins-clova&subdirectory=livekit-plugins/livekit-plugins-clova
pyyaml>=6.0.2
# livekit-plugins-turn-detector>=0.4.2 # Only available when using the VoicePipelineAgent?
